requirements.txt
# Core
pyyaml>=6.0
pydantic>=2.0
python-dotenv>=1.0

# LLM
openai>=1.0
anthropic>=0.20
httpx>=0.27       # 代理支持

# Financial Data
yfinance>=0.2.30
akshare>=1.12
tushare>=1.4.10   # A股备用数据源

# Data Processing
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3

# NLP & Dedup
sentence-transformers>=2.2
datasketch>=1.6

# Charts
matplotlib>=3.7
mplfinance>=0.12

# PDF Generation
reportlab>=4.0

# Web Scraping
requests>=2.31
beautifulsoup4>=4.12

# Async
aiohttp>=3.9
asyncio-throttle>=1.0
